+++ prepare/Learning05/Shape.cs
@@ -15,7 +15,7 @@ public class Shape
 
     }
     // I am unsure if the datatype needs to be float or int or something else.
-    public virtual float GetArea()
+    public virtual double GetArea()
     {
         // I am guessing here.
         return 0;
